+++ functions/package.json
@@ -21,6 +21,7 @@
     "firebase-functions": "^0.8.1",
     "firepad": "^1.4.0",
     "fs-extra": "^4.0.2",
+    "glob": "^7.1.2",
     "googleapis": "^23.0.0",
     "lodash": "^4.17.4",
     "mkdirp": "^0.5.1",
