+++ functions/src/storageFileToRTDB/index.js
